fix namespacing issues, some better error checking

+++ main.py
@@ -26,13 +26,16 @@ discover_functions_manager = DiscoverFunctionsManager(rate_limiter, rate_limiter
 discover_agents_manager = DiscoverAgentsManager(rate_limiter, rate_limiter_sync)
 functions_and_agents_metadata = FunctionsAndAgentsMetadata()
 
-agentcache = TTLCache(maxsize=16384, ttl=36000)
 discoverfunctioncache = TTLCache(maxsize=16384, ttl=36000)
 discoveragentscache = TTLCache(maxsize=16384, ttl=36000)
 
 @app.post('/discover_functions/')
 async def discoverFunctions(function_input: DiscoverFunctionsModel):
     """Endpoint to get functions based on provided input."""
+    if function_input.auth.api_key == '':
+        return {'response': "Error: LLM API key not provided", 'elapsed_time': 0}
+    if function_input.auth.namespace_id == '':
+        return {'response': "Error: namespace_id not provided", 'elapsed_time': 0}
     result = discoverfunctioncache.get(function_input)
     if result is not None:
         logging.info(f'Found functions in cache, result {result}')
@@ -51,6 +54,10 @@ async def discoverFunctions(function_input: DiscoverFunctionsModel):
 @app.post('/add_function/')
 async def addFunction(function_input: AddFunctionInput):
     """Endpoint to push functions based on provided functions."""
+    if function_input.auth.api_key == '':
+        return {'response': "Error: LLM API key not provided", 'elapsed_time': 0}
+    if function_input.auth.namespace_id == '':
+        return {'response': "Error: namespace_id not provided", 'elapsed_time': 0}
     logging.info(f'Adding function: {function_input.name}')
     functions = {}
     function_types = ['information_retrieval', 'communication', 'data_processing', 'sensory_perception', 'programming']
@@ -75,55 +82,62 @@ async def addFunction(function_input: AddFunctionInput):
 @app.post('/get_agent/')
 async def getAgent(agent_input: GetAgentModel):
     """Endpoint to get agent."""
-    agentcache
-    result = agentcache.get(agent_input.name)
-    if result is not None:
-        logging.info(f'Found agent in cache, result {result}')
-        return {'response': result, 'elapsed_time': 0}
+    if agent_input.auth.api_key == '':
+        return {'response': "Error: LLM API key not provided", 'elapsed_time': 0}
+    if agent_input.auth.namespace_id == '':
+        return {'response': "Error: namespace_id not provided", 'elapsed_time': 0}
+    if agent_input.name == '':
+        return {'response': "Error: agent name not provided!", 'elapsed_time': 0}
     response, elapsed_time = await functions_and_agents_metadata.get_agent(agent_input)
-    if len(response.name) > 0:
-        agentcache[agent_input.name] = response
-    # try to find it globally
-    else:
+    if len(response.name) == 0:
         agent_input.auth.namespace_id = ""
         response, elapsed_time = await functions_and_agents_metadata.get_agent(agent_input)
-        if len(response.name) > 0:
-            agentcache[agent_input.name] = response 
     return {'response': response, 'elapsed_time': elapsed_time}
 
 @app.post('/upsert_agent/')
 async def upsertAgent(agent_input: UpsertAgentInput):
     """Endpoint to upsert agent."""
+    if agent_input.auth.api_key == '':
+        return {'response': "Error: LLM API key not provided", 'elapsed_time': 0}
+    if agent_input.auth.namespace_id == '':
+        return {'response': "Error: namespace_id not provided", 'elapsed_time': 0}
+    if agent_input.name == '':
+        return {'response': "Error: agent name not provided!", 'elapsed_time': 0}
     logging.info(f'Upsert agent: {agent_input.name}')
     agents = {}
-    agent_types = ['information_retrieval', 'communication', 'data_processing', 'sensory_perception', 'programming', 'planning', 'groups', 'user']
-    if agent_input.category not in agent_types:
-        return {'response': f'Invalid category for agent {agent_input.name}, must be one of {agent_types}'}
-
-    # Append the new agent to the category
-    new_agent = {
-        'name': agent_input.name,
-        'description': agent_input.description
-    }
-    agents[agent_input.category] = [new_agent]
+    logging.info(f'Upsert agen0 {agent_input}')
+    if agent_input.category:
+        agent_types = ['information_retrieval', 'communication', 'data_processing', 'sensory_perception', 'programming', 'planning', 'groups', 'user']
+        if agent_input.category not in agent_types:
+            return {'response': f'Invalid category for agent {agent_input.name}, must be one of {agent_types}'}
+    logging.info('Upsert agen1')
     if agent_input.human_input_mode:
         human_input_types = ['ALWAYS', 'NEVER', 'TERMINATE']
         if agent_input.human_input_mode not in human_input_types:
             return {'response': f'Invalid human_input_mode for agent {agent_input.human_input_mode}, must be one of {human_input_types}'}
-
+    logging.info('Upsert agen2')
     # Push the agent
-    response, elapsed_time1 = await functions_and_agents_metadata.upsert_agent(agent_input)
+    response, elapsed_time1, agent = await functions_and_agents_metadata.upsert_agent(agent_input)
     if response != "success":
         return {'response': response, 'elapsed_time': elapsed_time1}
-    if agentcache.get(agent_input.name):
-        agentcache.pop(agent_input.name)
+    logging.info('Upsert agen3')
     discoveragentscache.clear()
+    # Append the new agent to the category
+    new_agent = {
+        'name': agent.name,
+        'description': agent.description
+    }
+    agents[agent.category] = [new_agent]
     result, elapsed_time2 = await discover_agents_manager.push_agents(agent_input.auth, agents)
     return {'response': result, 'elapsed_time': elapsed_time1+elapsed_time2}
 
 @app.post('/discover_agents/')
 async def discoverAgents(agent_input: DiscoverAgentsModel):
     """Endpoint to upsert an agent."""
+    if agent_input.auth.api_key == '':
+        return {'response': "Error: LLM API key not provided", 'elapsed_time': 0}
+    if agent_input.auth.namespace_id == '':
+        return {'response': "Error: namespace_id not provided", 'elapsed_time': 0}
     agent_types = ['information_retrieval', 'communication', 'data_processing', 'sensory_perception', 'programming', 'planning', 'groups', 'user']
 
     if agent_input.category not in agent_types:
@@ -142,13 +156,17 @@ async def discoverAgents(agent_input: DiscoverAgentsModel):
 @app.post('/delete_agent/')
 async def deleteAgent(agent_input: DeleteAgentModel):
     """Endpoint to delete agent."""
+    if agent_input.auth.api_key == '':
+        return {'response': "Error: LLM API key not provided", 'elapsed_time': 0}
+    if agent_input.auth.namespace_id == '':
+        return {'response': "Error: namespace_id not provided", 'elapsed_time': 0}
+    if agent_input.name == '':
+        return {'response': "Error: agent name not provided!", 'elapsed_time': 0}
     logging.info(f'Deleting agent: {agent_input.name}')
     # delete the agent
     response, elapsed_time1 = await functions_and_agents_metadata.delete_agent(agent_input)
     if response != "success":
         return {'response': response, 'elapsed_time': elapsed_time1}
-    if agentcache.get(agent_input.name):
-        agentcache.pop(agent_input.name)
     discoveragentscache.clear()
     result, elapsed_time2 = discover_agents_manager.delete_agent(agent_input.auth, agent_input.name)
     return {'response': result, 'elapsed_time': elapsed_time1+elapsed_time2}
